# public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\n\nHeader -> api_key :  \n\n",
    "version": "1.0.0",
    "title": "Solid Disk Direct"
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
...
  "paths": {
...
    "/contact_messages/store": {
      "post": {
        "tags": [
          "Contact Us"
        ],
        "summary": "Contact Us",
        "operationId": "contact-us",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Enter Name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Enter email",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "formData",
            "description": "Enter phone",
            "required": true,
            "type": "string"
          },
          {
            "name": "subject",
            "in": "formData",
            "description": "Enter subject",
            "required": true,
            "type": "string"
          },
          {
            "name": "message",
            "in": "formData",
            "description": "Enter message",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "api-key"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ]
}